fix: don't duplicate the first page of tables
Turns out the reason that BGA didn't like the page being set to 0 is
because there isn't a page 0! It starts from page one (and default to
one if not specified)

BGS has a bug where it doesn't de-duplicate between UUIDs that are
within the same file, so this was actually causing duplicate imports
despite the UUIDs.